+++ src/main/java/com/wayble/server/direction/service/TransportationService.java
@@ -4,10 +4,10 @@ import com.wayble.server.common.exception.ApplicationException;
 import com.wayble.server.direction.dto.TransportationGraphDto;
 import com.wayble.server.direction.dto.request.TransportationRequestDto;
 import com.wayble.server.direction.dto.response.TransportationResponseDto;
-import com.wayble.server.direction.entity.DirectionType;
 import com.wayble.server.direction.entity.transportation.Edge;
 import com.wayble.server.direction.entity.transportation.Node;
 import com.wayble.server.direction.entity.transportation.Route;
+import com.wayble.server.direction.entity.type.DirectionType;
 import com.wayble.server.direction.repository.EdgeRepository;
 import com.wayble.server.direction.repository.NodeRepository;
 import lombok.RequiredArgsConstructor;
@@ -689,7 +689,7 @@ public class TransportationService {
             } else if (currentType == DirectionType.SUBWAY) {
                 try {
                     if (currentEdge.getStartNode() != null) {
-                        TransportationResponseDto.NodeInfo nodeInfo = facilityService.getNodeInfo(currentEdge.getStartNode().getId(), currentEdge.getRoute().getRouteId());
+                        TransportationResponseDto.NodeInfo nodeInfo = facilityService.getNodeInfo(currentEdge.getStartNode().getId(), currentEdge.getRoute().getRouteId(), null);
                         
                         subwayInfo = new TransportationResponseDto.SubwayInfo(
                             nodeInfo.wheelchair(),
